doc: previews to backup CodeSandbox
I experience sporradic unavailability with the code sandbox MicroVM
embed that we're using in the book. It's very nice when it works, but I
was getting an error where I couldn't read the examples at all. This
drops the text of all MicroVM examples into the book itself behind a
`<preview>` so that readers can still view the source even if
CodeSandbox is wonky.