Fix Trunk not finding Rust project in "Getting Started"
As stated, the Getting Started directions would result in a blank page being served up. This is because Trunk's behavior is to look for Cargo.toml in the **parent** directory of the source HTML file https://trunkrs.dev/assets/#rust. In the provided instructions the source HTML file (index.html) and the Cargo.toml file are in the same directory, so Trunk would not find the Cargo.toml, serving up a blank page and emitting the warning "WARN no rust project found"

This behavior is not obvious as the homepage for Trunk suggests that the Cargo.toml and index.html can be in the same directory.