fix: handle cross-origin redirects in server function redirect hook
In client-side navigation we now handle redirects returned from
server functions the same way how the browser would handle them.
This fixes cross-origin redirects.

Note: this is a breaking change for relative locations such as
`Location: foo`. Previously that was resolved to `<origin>/foo`.
Now it's resolved as a URI-reference (RFC 3986) relative to the
server function URL's (typically `<origin>/api/something`),
resulting in `<origin>/api/foo`. This is required, because it's
the same behavior as with JS/WASM disabled. In case you want to
redirect to paths, make sure that it starts with a single slash
(e.g. `Location: /foo`).