feat: provide two different Url constructors: with and without base
The URL constructor without base (`Url::try_from`) requires an
URL, so we need to enforce this for all callers. The only caller
which called this with a path (i.e., without origin) was
`create_location()`, so I changed `create_location` to always
prepend `http://leptos.dev` to the path. The origin in this case
doesn't matter, because `create_location()` doesn't use the origin.

However, there was an inconsistency between client and server:
previosly, `create_location()` was called with a path on the client
and an URL (with origin `http://leptos.rs` or `http://leptos.dev`)
on the server. I fixed this by always using a path on the server,
too.