examples/cf-worker: Broken test investigation
Top of the tree is currently failing tests.
Failing on the following build

cd examples/cf-worker$
cargo build  --lib --release --target wasm32-unknown-unknown

* builds Ok without the --target specified
* builds Ok with --target specified BUT after cargo clean

so something is sticky, and messy crate dependency issue

My IDE reports "worker" and "worker-macros" are outdated
That is the first thing that needs to be resolved.