examples/cf-worker: Fixed intermittent CI issue?
I think this resolves CI bug where depending on the test execution
order a sub-library is build or not with the correct cfg flag.

The solution is to add a cfg conventionally added by the build.rs file
explicitly in a .cargo.config.toml

build.rs is run on workspace items. Things in the examples directory are not workspace items.