fix: Set content type if error occurs in Leptos middleware
Problem
-------
If an error occurs
[in Leptos's middleware chain](https://github.com/leptos-rs/leptos/blob/main/server_fn/src/middleware/mod.rs#L77),
the content-type header of the returned error response is not set.

Solution
--------
Save the error's content-type from the boxed service in
`BoxedService`, and pass the content type into `Service::run`.

Note that this requires the following breaking changes:
- Add a new field to `BoxedService`.
- Update the `Service::run` trait method to accept the error's content
  type as a parameter.

However, since this requires breaking changes anyway, this PR achieves
the following in a more extensible way (but still API-breaking):
- Replace the `ser` field in `BoxedService` with a `ServiceRunConfig`
  that contains the `ser` error serializer and the error's content
  type string.
- Mark `BoxedService` as `#[non_exhaustive]` so new fields can be added
  in the future without breaking the API.
- Replace the `ser` parameter of `Service::run` with the `ServiceRunConfig`
  This allows `Service::run` to remain the same even if
  new fields are added to the `ServiceRunConfig` in the future.