Add naive macro support for `MultipartData` codec
All we need to do is derive `Clone` and implement `FromFormData`. The
former is trivial. The later however is quite hacky, as it requires us
to 1. reference `leptos`, which is not necessarily there when just using
`server_fn`, and 2. makes the assumption that the only field is of type
`MultipartData`. I think that's a fair assumption though.